refactored CircleMenu methods, moved MainMenu into menu package.
