+++ Code/src/main/java/nl/utwente/group10/ui/components/blocks/Block.java
@@ -2,10 +2,10 @@ package nl.utwente.group10.ui.components.blocks;
 
 import java.io.IOException;
 
+import javafx.application.Platform;
 import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.StackPane;
-
 import nl.utwente.group10.haskell.expr.Expr;
 import nl.utwente.group10.ui.CustomUIPane;
 import nl.utwente.group10.ui.components.ComponentLoader;
@@ -34,7 +34,7 @@ public abstract class Block extends StackPane implements ComponentLoader {
     /** The pane that is used to hold state and place all components on. */
     private CustomUIPane parentPane;
     /** The context menu associated with this block instance. */
-    private CircleMenu contextMenu;
+    private CircleMenu circleMenu;
 
     /**
      * @param blockName
@@ -66,11 +66,11 @@ public abstract class Block extends StackPane implements ComponentLoader {
 
         this.addEventHandler(MouseEvent.MOUSE_CLICKED, this::handleMouseEvent);
 
-        this.createContextMenu();
+        Platform.runLater(() -> (this.createCircleMenu()));
     }
 
-    private void createContextMenu() {
-        contextMenu = new CircleMenu(this);
+    protected void createCircleMenu() {
+        circleMenu = new CircleMenu(this);
     }
 
     /** Returns the output Anchor for this Block. */
@@ -86,7 +86,7 @@ public abstract class Block extends StackPane implements ComponentLoader {
         if (parentPane.getSelectedBlock().isPresent()
                 && parentPane.getSelectedBlock().get().equals(this)
                 && t.getButton().equals(MouseButton.PRIMARY)) {
-            contextMenu.show(t);
+            circleMenu.show(t);
         } else {
             parentPane.setSelectedBlock(this);
         }
